ai-tools/abi-assistant/examples/gemini-agent/requirements.txt
# Requirements for Gemini Agent integration examples
aiohttp>=3.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0